+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Use minified bundle for CI testing
 - Improve featureIDBuffer generation by 20%
 - Trigger feature events (hover, enter, leave) when layers are updated
+- ramp($category, ...) implementation is fixed when using multiple categorical properties
 
 ### Removed
 - Removed XYZ expression